base_annotate/test_basic.sh: always specify the perf.data file
Originally, perf used to default to perf.data without '-i' or '-o' option,
however, now it seems to prefer stdio piping in the same situation.
However, the above is true only in "headless" mode, when perf is run by
automation. And thus, it causes problems in the testsuite.

This is an attempt to fix it in tests.