base_config/test_set.sh: test of setting variables to default values
This test parses all variable names from the help message and uses
'base_config/no_default.txt' file to set them to their default value. It
sets variables one by one and checks if they appear in --list option.
This test stores the previous configuration and restores it after the
test, but if the test was interrupted, it the config file will not be
restored. If this happens the previous configuration should still be in
'$CURRENT_TEST_DIR/.config_before' file.