base_record/test_intr_regs.sh: Fix 128bit registers handling
XMM registers are 128bit. Due to that, perf stores their values as a pair
of 64bit values in the trace. In the `perf report -D` dump, they appear
twice per sample, such as following:

  1113433709679 0x4180 [0x1d0]: PERF_RECORD_SAMPLE(IP, 0x4001): 69308/69308: 0xffffffffa6463f34 period: 1 addr: 0
  ... intr regs: mask 0xffffffff00ff0fff ABI 64-bit
  .... AX    0xff
  .... BX    0xffff965e9fe4f3e0
  .... CX    0x38f
  .... DX    0xf
  .... SI    0xff
  .... DI    0x38f
  .... BP    0xffff965c7d6d8060
  .... SP    0xffffad1bc51c7d00
  .... IP    0xffffffffa6463f34
  .... FLAGS 0x8
  .... CS    0x10
  .... SS    0x18
  .... R8    0x0
  .... R9    0x29580
  .... R10   0x2604025890d
  .... R11   0x0
  .... R12   0xffff965c7d6d8000
  .... R13   0xffff965e9fe71600
  .... R14   0xffff965c7d6d8000
  .... R15   0x1
  .... XMM0  0x0
  .... XMM0  0x0
  .... XMM1  0x0
  .... XMM1  0x0
  .... XMM2  0x0
  .... XMM2  0x0
  .... XMM3  0x0
  .... XMM3  0x0
  [...]

Due to that, the checker used to fail on machines with XMM registers.

To handle them correctly, the value of how many times such register
appeared in the log needs to be divided by two.

Also, move the check for samples out of the loop, it's enough to do it
once.