base_config/test_set_trace.sh: Do not require "[continued]" line in trace
The lines with "[continued]" are possible in the trace output, they are
completely correct, thus the related regexps are eligible for "all lines
matched" checks. However, they aren't compulsory, thus they should not
appear in "all patterns found" checks.

The "continued" lines appear in trace, in case some syscall takes too long
and thus its record needs two rows in the trace output -- when it starts
and when it finishes. Since the traced workload is arbitrary, it does not
have to contain such syscalls.