base_probe/test_advanced.sh: Allow multiple kprobes
Sometimes when adding a kprobe by perf, it results in multiple probe
points, such as the following:

  # perf probe -a 'do_sys_open filename:string'
  Added new events:
    probe:do_sys_open    (on do_sys_open with filename:string)
    probe:do_sys_open_1  (on do_sys_open with filename:string)

In case the second one is actually hit, the test used to fail, because
it didn't use all of them for recording and also, it didn't expect the
"_1" suffix in the logs.

Currently, it seems that perf is capable of probing both probe points
with the string argument (which used to be broken). So, we just need
to fix the test.