base_timechart/test_io.sh: test of io options of perf timechart tool
At first we record data with 'perf timechart record' with --io-only
option and then we create svg files with all io options, which are
--io-skip-eagain, --io-min-time and --io-merge-dist. There are the same
tests on svf file output as in 'base_timechart/test_basic.sh' file, but
it is better to have these files separated. In this file we also check
if --io-min-time works correctly with <title> element merges text.