common/check_timestamps.pl: generalize timestamps check
The old version of this file could test only specific input, because the
pattern we used to check timestamps was embedded in the file. This
version takes the first argument of 'common/check_timestamps.pl' as the
pattern, so we can now test order of any timestamps or if needed, order
of something totally different.