base_script/test_time.sh: test for time options of perf script command
In this test, there are checks for these options: --ns, --time,
--deltatime and --reltime. Each option is tested for expected output and
also for the number of samples, which should match the number reported
by the 'perf record' command. Options --ns, --time and --reltime are
also tested for correct timestamp order.

In the --ns option test, we check whether the timestamps have nine
decimal digits.
There are file comparisons in the --time option test, where we check if
--time 0%-10% gives us the same output as --time 10%/1 and so on.
There is also a test, which checks if times in the --deltatime option
were counted correctly.