base_script/subtest__stackcollapse.sh: changed output check
Instead of checking whether 'perf' and 'sleep' commands are in the
output of 'perf script stackcollapse report', we now compare the output
of the command with the expected output. The expected output is created
by a series of commands and then compared by diff command to 'perf
script stackcollapse report' output.