base_script/test_field_tod.sh: test of perf script tod field
This test checks whether 'perf script -F+tod' works correctly. Before
the testing, we need to record the data with the clockid specified. The
'perf script -F+tod' should display the date alongside with time of the
day.

There are tests that check whether the first date and time and the last
date and time are within the expected bounds. There is also a test, in
which we check if the difference between two consecutive dates and times
is approximately the same as the difference between two consecutive
timestamps.